Initial Push
I have it broken into the main bulk Superheterodyne note, which I want to include links and the various frequencies we need. Then I have a subfolder with each component. I am going to include each DigiKey component I have in there. My goal for the future is to have notes written for how to simulate using Altium, Ansys, and ADS(maybe) so we are both working off of the same information